库存四位数的变更

+++ src/main/java/com/aiqin/bms/scmp/api/product/domain/response/QueryStockSkuListRespVo.java
@@ -6,6 +6,7 @@ import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -96,7 +97,7 @@ public class QueryStockSkuListRespVo implements Serializable {
 
     @ApiModelProperty("昨天含税成本")
     @JsonProperty("tax_cost")
-    private Long taxCost;
+    private BigDecimal taxCost;
 
     @ApiModelProperty("批次号")
     @JsonProperty(value = "batch_code")
